- Split client App.js to components based on form submit - No major code change (just splitting)
